BitKeeper/etc/logging_ok:
  auto-union
configure.in:
  Auto merged
mysql-test/r/ps_6bdb.result:
  Auto merged
mysql-test/r/ps_7ndb.result:
  Auto merged
sql/field.cc:
  Auto merged
sql/field.h:
  Auto merged
sql/log_event.cc:
  Auto merged
sql/mysql_priv.h:
  Auto merged
sql/sql_acl.cc:
  Auto merged
sql/sql_select.cc:
  Auto merged
sql/sql_table.cc:
  Auto merged
sql/sql_yacc.yy:
  Auto merged
storage/myisam/mi_check.c:
  Auto merged
storage/ndb/src/kernel/blocks/ndbfs/Ndbfs.cpp:
  Auto merged
storage/ndb/test/run-test/ndb-autotest.sh:
  Auto merged

+++ storage/ndb/test/run-test/ndb-autotest.sh
@@ -13,7 +13,7 @@ save_args=$*
 VERSION="ndb-autotest.sh version 1.04"
 
 DATE=`date '+%Y-%m-%d'`
-HOST=`hostname -s`
+HOST=`hostname`
 export DATE HOST
 
 set -e
@@ -330,7 +330,10 @@ start(){
 		tar cfz /tmp/res.$2.$$.tgz `basename $p2`/$DATE
 		scp /tmp/res.$2.$$.tgz \
 		    $result_host:$result_path/res.$DATE.$HOST.$2.$$.tgz
-		rm -f /tmp/res.$2.$$.tgz
+		if [ $? -eq 0 ]
+		then
+		    rm -f /tmp/res.$2.$$.tgz
+		fi
 	fi
 }
